Renaming from feature map to encoding map (step from classical to quantum)
